requirements.txt
Flask>=2.0
openai>=1.0.0
python-dotenv>=0.21.0
black==23.9.0
isort==5.12.0
mypy==1.8.0


# optional: for better code diff display
diff-match-patch==20200713